+++ packages/rma-server/src/warranty-claim/policies/warranty-claim-policies/warranty-claim-policies.service.ts
@@ -103,7 +103,7 @@ export class WarrantyClaimPoliciesService {
     );
   }
 
-  validateWarrantySerailNo(claimsPayload: WarrantyClaimDto) {
+  validateWarrantySerialNo(claimsPayload: WarrantyClaimDto) {
     return from(
       this.serialNoService.findOne({
         serial_no: claimsPayload.serial_no,