+++ packages/rma-warranty/src/app/warranty-ui/add-warranty-claim/add-warranty-claim.page.ts
@@ -22,15 +22,14 @@ import {
 import {
   SOMETHING_WENT_WRONG,
   ITEM_BRAND_FETCH_ERROR,
-  SERIAL_FETCH_ERROR,
   USER_SAVE_ITEM_SUGGESTION,
   ITEM_NOT_FOUND,
 } from '../../constants/messages';
 import { ActivatedRoute, Router } from '@angular/router';
 import { AUTH_SERVER_URL, TIME_ZONE } from '../../constants/storage';
 import { DateTime } from 'luxon';
-import { StorageService } from '../../api/storage/storage.service';
 import { WarrantyService } from '../warranty-tabs/warranty.service';
+
 @Component({
   selector: 'app-add-warranty-claim',
   templateUrl: './add-warranty-claim.page.html',
@@ -86,12 +85,11 @@ export class AddWarrantyClaimPage implements OnInit {
   constructor(
     private location: Location,
     private readonly time: TimeService,
-    private readonly warrantyService: AddWarrantyService,
+    private readonly addWarrantyService: AddWarrantyService,
     private readonly loadingController: LoadingController,
     private readonly snackbar: MatSnackBar,
     private readonly router: Router,
     private readonly activatedRoute: ActivatedRoute,
-    private readonly storage: StorageService,
     private readonly warrantyClaim: WarrantyService,
   ) {}
 
@@ -119,26 +117,21 @@ export class AddWarrantyClaimPage implements OnInit {
       category: { disabled: true, active: true },
     };
     this.createForm();
-    this.warrantyClaimForm.controls.received_on.setValue(
-      await (await this.getDateTime(new Date())).date,
-    );
-    this.warrantyClaimForm.controls.delivery_date.setValue(
-      await this.getDeliveryDate(new Date()),
-    );
+    this.setDefaults();
 
     this.filteredCustomerList = this.warrantyClaimForm.controls.customer_name.valueChanges.pipe(
       debounceTime(500),
       startWith(''),
-      this.warrantyService.getCustomerList(),
+      this.addWarrantyService.getRelayedCustomerList(),
     );
 
     this.productList = this.warrantyClaimForm.controls.product_name.valueChanges.pipe(
       debounceTime(500),
       startWith(''),
-      this.warrantyService.getItemList(),
+      this.addWarrantyService.getItemList(),
     );
 
-    this.warrantyService
+    this.addWarrantyService
       .getStorage()
       .getItem('territory')
       .then(territory => {
@@ -150,7 +143,7 @@ export class AddWarrantyClaimPage implements OnInit {
       debounceTime(500),
       startWith(''),
       switchMap(value => {
-        return this.warrantyService.getProblemList(value);
+        return this.addWarrantyService.getProblemList(value);
       }),
       map(res => res.docs),
     );
@@ -158,6 +151,17 @@ export class AddWarrantyClaimPage implements OnInit {
       this.clearAllControlValidators();
       this.setValues();
     }
+  }
+  async setDefaults() {
+    this.warrantyClaimForm.controls.received_on.setValue(
+      await (await this.getDateTime(new Date())).date,
+    );
+    this.warrantyClaimForm.controls.delivery_date.setValue(
+      await this.getDeliveryDate(new Date()),
+    );
+    this.warrantyClaimForm.controls.claim_type.setValue(this.claimList[0]);
+    this.warrantyClaimForm.controls.category.setValue(CATEGORY.SINGLE);
+    this.getFormState(this.warrantyClaimForm.controls.claim_type.value);
   }
 
   clearAllControlValidators() {
@@ -248,7 +252,7 @@ export class AddWarrantyClaimPage implements OnInit {
     const loading = await this.loadingController.create();
     await loading.present();
     const payload = this.mapUpdateClaim();
-    this.warrantyService.updateWarrantyClaim(payload).subscribe({
+    this.addWarrantyService.updateWarrantyClaim(payload).subscribe({
       next: () => {
         loading.dismiss();
         this.router.navigate(['/warranty']);
@@ -382,24 +386,33 @@ export class AddWarrantyClaimPage implements OnInit {
   }
 
   async createClaim() {
+    if (!this.warrantyClaimForm.valid) {
+      this.snackbar.open('Please enter missing fields.', CLOSE, {
+        duration: 4500,
+      });
+      this.warrantyClaimForm.markAllAsTouched();
+      return;
+    }
     const loading = await this.loadingController.create();
     await loading.present();
     const detail = await this.assignFields();
     if (this.warrantyClaimForm.controls.category.value === CATEGORY.BULK) {
       if (detail.bulk_products?.length > 1) {
-        return this.warrantyService.createBulkWarrantyClaim(detail).subscribe({
-          next: () => {
-            loading.dismiss();
-            this.router.navigate(['/warranty']);
-          },
-          error: ({ message }) => {
-            loading.dismiss();
-            if (!message) message = SOMETHING_WENT_WRONG;
-            this.snackbar.open(message, 'Close', {
-              duration: DURATION,
-            });
-          },
-        });
+        return this.addWarrantyService
+          .createBulkWarrantyClaim(detail)
+          .subscribe({
+            next: () => {
+              loading.dismiss();
+              this.router.navigate(['/warranty']);
+            },
+            error: ({ message }) => {
+              loading.dismiss();
+              if (!message) message = SOMETHING_WENT_WRONG;
+              this.snackbar.open(message, 'Close', {
+                duration: DURATION,
+              });
+            },
+          });
       }
       loading.dismiss();
       this.snackbar.open(`Please use single claim for one product`, CLOSE, {
@@ -407,17 +420,14 @@ export class AddWarrantyClaimPage implements OnInit {
       });
     }
 
-    return this.warrantyService.createWarrantyClaim(detail).subscribe({
+    return this.addWarrantyService.createWarrantyClaim(detail).subscribe({
       next: () => {
         loading.dismiss();
         this.router.navigate(['/warranty']);
       },
-      error: ({ message }) => {
+      error: error => {
         loading.dismiss();
-        if (!message) message = SOMETHING_WENT_WRONG;
-        this.snackbar.open(message, 'Close', {
-          duration: DURATION,
-        });
+        this.getMessage(error?.error?.message || 'Error in creating claim.');
       },
     });
   }
@@ -487,10 +497,16 @@ export class AddWarrantyClaimPage implements OnInit {
   async customerChanged(customer) {
     const loading = await this.loadingController.create();
     await loading.present();
-    this.warrantyService.getAddress(customer.name).subscribe({
-      next: res => {
+    return this.addWarrantyService.getRelayCustomer(customer.name).subscribe({
+      next: (res: any) => {
+        // this clearly needs rework. dont try to optimize. rewrite it.
         loading.dismiss();
         this.contact = res;
+        if (!res) {
+          this.getMessage('Failed to fetch customer.');
+          return;
+        }
+        this.warrantyClaimForm.controls.customer_name.setValue(res);
         if (!res.customer_primary_address) {
           if (!res.mobile_no) {
             this.snackbar.open(
@@ -500,6 +516,8 @@ export class AddWarrantyClaimPage implements OnInit {
                 duration: DURATION,
               },
             );
+            this.warrantyClaimForm.controls.customer_contact.setValue('');
+            this.warrantyClaimForm.controls.customer_address.setValue('');
           } else {
             this.snackbar.open('Address Not found', 'Close', {
               duration: DURATION,
@@ -524,6 +542,14 @@ export class AddWarrantyClaimPage implements OnInit {
           );
         }
       },
+      error: err => {
+        loading.dismiss();
+        this.snackbar.open(
+          err?.error?.message || 'Error fetching customer',
+          CLOSE,
+          { duration: 4200 },
+        );
+      },
     });
   }
   getOptionText(option) {
@@ -547,24 +573,20 @@ export class AddWarrantyClaimPage implements OnInit {
   }
 
   async serialChanged(name) {
-    const timeZone = await this.storage.getItem(TIME_ZONE);
-    this.warrantyService.getSerial(name).subscribe({
+    const timeZone = await this.addWarrantyService
+      .getStorage()
+      .getItem(TIME_ZONE);
+    this.addWarrantyService.getSerial(name).subscribe({
       next: (res: SerialNoDetails) => {
         this.getSerialData = res;
         if (res.claim_no) {
-          this.snackbar.open(
-            `Claim already exists serial no ${res.serial_no}`,
-            'Close',
-            {
-              duration: DURATION,
-            },
-          );
+          this.getMessage(`Claim already exists serial no ${res.serial_no}`);
           return;
         }
-        if (!res.delivery_note) {
-          this.snackbar.open('Serial not sold yet', 'Close', {
-            duration: DURATION,
-          });
+        if (!res.customer) {
+          this.getMessage(
+            'Serial not sold or serials is not linked to customer.',
+          );
           return;
         }
         if (
@@ -601,12 +623,21 @@ export class AddWarrantyClaimPage implements OnInit {
         this.itemOptionChanged({ item_code: res.item_code });
         this.customerChanged({ name: res.customer });
       },
-      error: ({ message }) => {
-        if (!message) message = `${SOMETHING_WENT_WRONG}${SERIAL_FETCH_ERROR}`;
-        this.snackbar.open(message, 'Close', {
-          duration: DURATION,
-        });
+      error: error => {
+        this.snackbar.open(
+          error?.error?.message || 'Error Fetching provided serial.',
+          'Close',
+          {
+            duration: DURATION,
+          },
+        );
       },
+    });
+  }
+
+  getMessage(message) {
+    this.snackbar.open(message, CLOSE, {
+      duration: DURATION,
     });
   }
 
@@ -625,7 +656,7 @@ export class AddWarrantyClaimPage implements OnInit {
   }
 
   itemOptionChanged(option) {
-    this.warrantyService.getItem(option.item_code).subscribe({
+    this.addWarrantyService.getItem(option.item_code).subscribe({
       next: (res: WarrantyItem) => {
         this.itemDetail = res;
         if (!res.brand) {
@@ -643,7 +674,7 @@ export class AddWarrantyClaimPage implements OnInit {
   }
 
   getItemBrandFromERP(item_code: string) {
-    this.warrantyService.getItemBrandFromERP(item_code).subscribe({
+    this.addWarrantyService.getItemBrandFromERP(item_code).subscribe({
       next: res => {
         if (!res.brand) {
           this.snackbar.open(ITEM_BRAND_FETCH_ERROR, 'Close', {
@@ -668,7 +699,7 @@ export class AddWarrantyClaimPage implements OnInit {
   }
 
   openERPItem(item_code: string) {
-    this.warrantyService
+    this.addWarrantyService
       .getStorage()
       .getItem(AUTH_SERVER_URL)
       .then(auth_url => {