+++ packages/rma-warranty/src/app/warranty-ui/add-warranty-claim/add-warranty.service.spec.ts
@@ -1,7 +1,10 @@
 import { TestBed } from '@angular/core/testing';
 
 import { AddWarrantyService } from './add-warranty.service';
-import { STORAGE_TOKEN } from '../../api/storage/storage.service';
+import {
+  StorageService,
+  STORAGE_TOKEN,
+} from '../../api/storage/storage.service';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 describe('AddWarrantyService', () => {
@@ -10,7 +13,16 @@ describe('AddWarrantyService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [HttpClientTestingModule],
-      providers: [{ provide: STORAGE_TOKEN, useValue: {} }],
+      providers: [
+        { provide: STORAGE_TOKEN, useValue: {} },
+        {
+          provide: StorageService,
+          useValue: {
+            getItem: (...args) => Promise.resolve('ITEM'),
+            getItems: (...args) => Promise.resolve([]),
+          },
+        },
+      ],
     });
     service = TestBed.inject(AddWarrantyService);
   });